[PRODDEV-431] Refactor listings to have the same behavior everywhere. Hide "More" buttons if there no extra content except the one already shown in the listing.
