+++ js/gated-content/src/mixins/ListingMixin.js
@@ -2,15 +2,42 @@ import { SettingsMixin } from '@/mixins/SettingsMixin';
 
 export const ListingMixin = {
   mixins: [SettingsMixin],
+  props: {
+    pagination: {
+      type: Boolean,
+      default: false,
+    },
+    limit: {
+      type: Number,
+      default: 0,
+    },
+  },
   data() {
     return {
       component: null,
       listing: [],
+      links: {},
     };
   },
   computed: {
     listingIsNotEmpty() {
       return this.listing.length > 0;
+    },
+    getPageParam() {
+      const pageParam = {
+        limit: 50,
+      };
+      if (this.pagination) {
+        const currentPage = parseInt(this.$route.query.page, 10) || 0;
+        pageParam.limit = this.config.pager_limit;
+        pageParam.offset = currentPage * pageParam.limit;
+      } else if (this.limit !== 0) {
+        pageParam.limit = this.limit;
+      }
+      return pageParam;
+    },
+    hasMoreItems() {
+      return typeof this.links.next !== 'undefined';
     },
     layoutClass() {
       if (this.component && this.config.components[this.component].show_covers) {
